Add basic GDPR consent placeholder form that will set flags accordingly
Provided a GDPR Singleton class to track the state. When the user wishes the edit / change the status, they'll edit just the singleton to prevenet any confusion
Added a StateGenController temp view for the next scope of work
Renamed the VCSingleton class to remove unecessary dashes
Modified main storyboard view to have a temporary red edit button on the MainViewcontroller to trigger GDPR editable state
Created a GDPRPopover view which has placeholder status items and 2 segues that will trigger back to main tab view
On load, the GDPRPopover is the entry point. If the GDPR status is true it will load the TabView right away, if not it will prompt user to select